Problem
When running an application/benchmark on the same machine than JITWatch,
it's unconvenient to have to specify the location of the all the jars that
you just run. (The information is already present in the log file.)

Solution

Use the information found in the hotspot log, and dynamically add new
urls to the DisposableClassLoader.

Extract the path of the jar from line like this:
[Loaded java.lang.ClassLoader from /Library/Java/JavaVirtualMachines/jdk1.7.0_65.jdk/Contents/Home/jre/lib/rt.jar]

Result

In case you run the code on the same machine than JITWatch -> Less configuration.
In the other case -> no changes.